feat: use 1st account colors on login & portfolio
